// web-app/customer/schemas/api-types.ts
export interface components {
  schemas: {
// ...
    Availability: {
      id?: number;
      staff_id?: number;
      practice_id?: number;
      team_member_id: number;
      /** Format: date-time */
      start_time: string;
      /** Format: date-time */
      end_time: string;
      /** Format: date-time */
      schedule_release_time?: string;
      team_member?: {
        readonly id?: number;
        readonly first_name: string;
        readonly last_name: string;
        readonly job_title: string;
        readonly bio?: string;
        /** Format: date-time */
        readonly created_at?: string;
        /** Format: date-time */
        readonly updated_at?: string;
      };
      /** Format: date-time */
      created_at?: string;
      /** Format: date-time */
      updated_at?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}